1. 推进日志的配置热更改造 2. kernel支持设置flags, 避免引入其他项目的时候导致对项目的干扰

+++ kernel/source/comp/Poller/Poller.cpp
@@ -622,7 +622,7 @@ void Poller::SafeEventLoop()
                     if(firstExpiredNano > 0)
                     {
                         firstExpiredNano /= KERNEL_NS::TimeDefs::NANO_SECOND_PER_MILLI_SECOND;
-                        _eventGuard.TimeWait(firstExpiredNano  > maxSleepMilliseconds ? firstExpiredNano : maxSleepMilliseconds);
+                        _eventGuard.TimeWait(static_cast<UInt64>(firstExpiredNano)  > maxSleepMilliseconds ? static_cast<UInt64>(firstExpiredNano) : maxSleepMilliseconds);
                     }
                     else
                     {
